pyproject.toml
[project]
name = "simple-fastmcp-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["boto3","botocore","mcp", "mcp-server","mcp-client","python-dotenv", "aioboto3"]
